Fix site breakage related to the Tracker-Tally CSS (#727)
* On Firefox, the Tracker-Tally can break sites that run JavaScript
that interacts with the injected css code. A recent example is
"https://mycovidrecord.health.nz/", which fails with:

Uncaught DOMException: CSSStyleSheet.cssRules getter: Not allowed to access cross-origin stylesheet

These older examples could be duplicates as well:
* https://github.com/niklasvh/html2canvas/issues/2197
* https://stackoverflow.com/a/62475347/783510

We can delay the css injection to "document_end" instead of
"document_start". It may still happen, but it should make it
less likely.

* Make the Tracker-Tally opt-in for new installations.

* Remove dead variable