Fixes a race in observe.
Background: observers (in the Set) and the normal wrapper call
when calling "observe" race for being the first to see the property
change. The second call will no longer block.

This change changes the semantic to provide the guarantee to the caller
of observe that the wrapper function finished its first call.