Include also the subdomain information when doing a TrackerDB lookup.
For instance, the request to
https://global.ketchcdn.com/web/v3/config/avant/website_smart_tag/boot.js
has
hostname=global.ketchcdn.com and domain=ketchcdn.com

The current TrackerDB entry is for "global.ketchcdn.com", so the lookup
by domain does not return a result. The adblocker implementation will
still do a lookup for the parent domain. So, it will also try out
"ketchcdn.com". Because of that, there should be no need to fallback
to the (parent) domain in the extension.